Refactor File Metadata Post processing
@patdunlavey can you test? Sorry for the lateness. Needed to super test. Thanks!
