+++ config/schema/strawberryfield.schema.yml
@@ -189,6 +189,9 @@ strawberryfield.filepersister_service_settings:
     mediainfo_exec_path:
       type: string
       label: 'mediainfo binary full executable path'
+    delete_tempfiles:
+      type: boolean
+      label: 'deleted temporary files generated by file metadata extraction immediately'
 
 strawberryfield.general:
   type: config_object
