+++ strawberryfield.services.yml
@@ -2,9 +2,14 @@ services:
   strawberryfield.utility:
     class: Drupal\strawberryfield\StrawberryfieldUtilityService
     arguments: [ '@file_system', '@entity_type.manager', '@config.factory','@module_handler', '@entity_field.manager', '@plugin.manager.search_api.parse_mode']
+  strawberryfield.file_metadata_extractor:
+    class: Drupal\strawberryfield\StrawberryfieldFileMetadataService
+    arguments: [ '@file_system', '@stream_wrapper_manager', '@config.factory', '@logger.factory', '@strawberryfield.utility' ]
+    tags:
+      - { name: backend_overridable }
   strawberryfield.file_persister:
     class: Drupal\strawberryfield\StrawberryfieldFilePersisterService
-    arguments: [ '@file_system', '@file.usage', '@entity_type.manager', '@stream_wrapper_manager', '@plugin.manager.archiver', '@config.factory', '@current_user', '@language_manager', '@transliteration', '@module_handler', '@logger.factory', '@strawberryfield.utility']
+    arguments: [ '@file_system', '@file.usage', '@entity_type.manager', '@stream_wrapper_manager', '@plugin.manager.archiver', '@config.factory', '@current_user', '@language_manager', '@transliteration', '@module_handler', '@logger.factory', '@strawberryfield.utility', '@strawberryfield.file_metadata_extractor']
     tags:
       - { name: backend_overridable }
   strawberryfield.keyname_manager:
